pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-army"
version = "0.1.0"
description = "A secure multi-agent AI system with orchestration, messaging, and task management"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "AgentArmy Contributors", email = "team@agentarmy.dev"}
]
keywords = ["ai", "agents", "multi-agent", "orchestration", "asyncio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Internet",
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pydantic-ai>=0.1.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "pyyaml>=6.0.0",
    # CLI terminal
    "rich>=13.0.0",
    # Message queues & cache
    "redis>=5.0.0",
    "aio-pika>=9.0.0",
    # Database
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    # Knowledge graph
    "neo4j>=5.0.0",
    "hypernetx>=2.3.0",
    # LLM clients
    "anthropic>=0.19.0",
    "ollama>=0.1.0",
    # MCP protocol
    "mcp>=1.0.0",
    # Security
    "cryptography>=41.0.0",
    "pyjwt>=2.8.0",
    "python-jose[cryptography]>=3.3.0",
    # Voice (TTS/STT)
    "edge-tts>=6.1.0",
    "faster-whisper>=1.0.0",
    # Desktop app
    "pywebview>=5.0",
]

[project.optional-dependencies]
desktop = [
    "pywebview>=5.0",
    "Pillow>=10.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "isort>=5.13.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
agentarmy = "scripts.cli:main"

[project.urls]
Homepage = "https://github.com/agentarmy/agentarmy"
Documentation = "https://agentarmy.dev"
Repository = "https://github.com/agentarmy/agentarmy.git"
Issues = "https://github.com/agentarmy/agentarmy/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
src = ["py.typed", "dashboard/*.html"]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "W", "I", "N", "UP", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]
